Gestion de inventario con funciones de verificar stock, descontar stock, listar productos y guardar productos, posee seguridad, aun falta enlazarla con otros microservicios de ser necesario, revisar seguridad

+++ src/main/java/cl/ecomarket/api/repository/ProductoRepository.java
@@ -3,4 +3,6 @@ package cl.ecomarket.api.repository;
 import cl.ecomarket.api.model.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductoRepository extends JpaRepository<Producto, Integer> {}
+public interface ProductoRepository extends JpaRepository<Producto, Long> {
+
+}
